patch: refactor new_gradient as GradientData trait impl
